რედაქტირებული შეკვეთის ცნება და წაშლილის ეგრევე ასახვა

+++ app/src/main/java/com/example/beerdistrkt/fragPages/orders/view/OrderView.kt
@@ -3,7 +3,6 @@ package com.example.beerdistrkt.fragPages.orders.view
 import android.content.Context
 import android.content.res.ColorStateList
 import android.graphics.Color
-import android.os.Build
 import android.util.AttributeSet
 import android.view.View
 import android.view.ViewGroup
@@ -13,6 +12,7 @@ import com.example.beerdistrkt.R
 import com.example.beerdistrkt.fragPages.orders.adapter.OrderItemAdapter
 import com.example.beerdistrkt.models.Order
 import com.example.beerdistrkt.models.OrderStatus
+import com.example.beerdistrkt.showToast
 import com.example.beerdistrkt.utils.visibleIf
 import kotlinx.android.synthetic.main.view_order.view.*
 
@@ -30,6 +30,8 @@ class OrderView @JvmOverloads constructor(
         View.inflate(context, R.layout.view_order, this)
         layoutParams =
             LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT)
+
+        orderUnitHistoryImg.setOnClickListener(this)
     }
 
     private fun resetForm() {
@@ -49,6 +51,7 @@ class OrderView @JvmOverloads constructor(
     fun fillData(order: Order) {
         resetForm()
         orderUnitClientNameTv.text = order.client.dasaxeleba
+        orderUnitHistoryImg.visibleIf(order.isEdited > 0)
         orderUnitCommentImg.visibleIf(!order.comment.isNullOrEmpty())
         orderUnitCheckImg.visibleIf(order.items.any { it.check == 1 })
         if (order.needCleaning == 1) {
@@ -91,7 +94,12 @@ class OrderView @JvmOverloads constructor(
             order.onEditClick()
         }
         orderUnitDeleteBtn.setOnClickListener {
-            order.onDeleteClick()
+            if (order.orderStatus != OrderStatus.DELETED)
+                order.onDeleteClick()
+            else {
+                context.showToast(R.string.deleted)
+                orderUnitRootSwipe.close(true)
+            }
         }
         orderUnitChangeDistributorBtn.setOnClickListener {
             order.onChangeDistributorClick()
@@ -110,6 +118,8 @@ class OrderView @JvmOverloads constructor(
     }
 
     override fun onClick(v: View?) {
-
+        when (v?.id) {
+            R.id.orderUnitHistoryImg -> context.showToast("show hist, TO DO")
+        }
     }
 }
